Turn EnsoDateTimeFormatter into an API facade
