+++ distribution/lib/Standard/Base/0.0.0-dev/src/Data/Time/Date_Time_Formatter.enso
@@ -139,7 +139,7 @@ type Date_Time_Formatter
             analyzer = Analyzer.new parsed
             java_formatter = analyzer.validate_after_parsing <|
                 As_Java_Formatter_Interpreter.interpret used_locale parsed
-            Date_Time_Formatter.Value (EnsoDateTimeFormatter.new java_formatter pattern FormatterKind.SIMPLE) deferred_parsing_warnings=analyzer.get_parsing_only_warnings
+            Date_Time_Formatter.Value (EnsoDateTimeFormatter.create java_formatter pattern FormatterKind.SIMPLE) deferred_parsing_warnings=analyzer.get_parsing_only_warnings
 
     ## ---
        icon: convert
@@ -193,7 +193,7 @@ type Date_Time_Formatter
             analyzer = Analyzer.new parsed
             java_formatter = analyzer.validate_after_parsing <|
                 As_Java_Formatter_Interpreter.interpret used_locale parsed
-            Date_Time_Formatter.Value (EnsoDateTimeFormatter.new java_formatter pattern FormatterKind.ISO_WEEK_DATE) deferred_parsing_warnings=analyzer.get_parsing_only_warnings
+            Date_Time_Formatter.Value (EnsoDateTimeFormatter.create java_formatter pattern FormatterKind.ISO_WEEK_DATE) deferred_parsing_warnings=analyzer.get_parsing_only_warnings
 
     ## ---
        advanced: true
@@ -221,7 +221,7 @@ type Date_Time_Formatter
             used_locale = if locale == Locale.default then Locale.us else locale
             java_locale = (used_locale.if_nothing Locale.us).java_locale
             java_formatter = DateTimeFormatter.ofPattern pattern java_locale
-            Date_Time_Formatter.Value (EnsoDateTimeFormatter.new java_formatter pattern FormatterKind.RAW_JAVA)
+            Date_Time_Formatter.Value (EnsoDateTimeFormatter.create java_formatter pattern FormatterKind.RAW_JAVA)
         _ -> Error.throw (Illegal_Argument.Error "The pattern must either be a string or a Java DateTimeFormatter instance.")
 
     ## ---
